Final/requirements.txt
flask
flask_restx
pyYAML
dotenv
gradio
langchain
langchain_community
langchain_openai


